+++ core/repository/paging/src/main/java/org/mozilla/social/core/repository/paging/FederatedTimelineRemoteMediator.kt
@@ -1,11 +1,10 @@
-package org.mozilla.social.feed.remoteMediators
+package org.mozilla.social.core.repository.paging
 
 import androidx.paging.ExperimentalPagingApi
 import androidx.paging.LoadType
 import androidx.paging.PagingState
 import androidx.paging.RemoteMediator
 import org.mozilla.social.core.database.model.entities.statusCollections.FederatedTimelineStatusWrapper
-import org.mozilla.social.core.usecase.mastodon.timeline.RefreshFederatedTimeline
 
 @OptIn(ExperimentalPagingApi::class)
 class FederatedTimelineRemoteMediator(
