+++ core/repository/paging/src/main/java/org/mozilla/social/core/repository/paging/LocalTimelineRemoteMediator.kt
@@ -1,11 +1,10 @@
-package org.mozilla.social.feed.remoteMediators
+package org.mozilla.social.core.repository.paging
 
 import androidx.paging.ExperimentalPagingApi
 import androidx.paging.LoadType
 import androidx.paging.PagingState
 import androidx.paging.RemoteMediator
 import org.mozilla.social.core.database.model.entities.statusCollections.LocalTimelineStatusWrapper
-import org.mozilla.social.core.usecase.mastodon.timeline.RefreshLocalTimeline
 
 @OptIn(ExperimentalPagingApi::class)
 class LocalTimelineRemoteMediator(
