+++ core/repository/paging/src/main/java/org/mozilla/social/core/repository/paging/MutesListRemoteMediator.kt
@@ -1,35 +1,32 @@
-package org.mozilla.social.core.usecase.mastodon.remotemediators
+package org.mozilla.social.core.repository.paging
 
 import androidx.paging.ExperimentalPagingApi
 import androidx.paging.LoadType
 import androidx.paging.PagingState
 import androidx.paging.RemoteMediator
 import kotlinx.coroutines.delay
-import org.mozilla.social.common.Rel
-import org.mozilla.social.common.getMaxIdValue
-import org.mozilla.social.core.database.model.entities.accountCollections.Follower
-import org.mozilla.social.core.database.model.entities.accountCollections.FollowerWrapper
+import org.mozilla.social.core.database.model.entities.accountCollections.MuteWrapper
 import org.mozilla.social.core.repository.mastodon.AccountRepository
 import org.mozilla.social.core.repository.mastodon.DatabaseDelegate
-import org.mozilla.social.core.repository.mastodon.FollowersRepository
+import org.mozilla.social.core.repository.mastodon.MutesRepository
 import org.mozilla.social.core.repository.mastodon.RelationshipRepository
+import org.mozilla.social.core.repository.mastodon.model.status.toDatabaseMute
 import timber.log.Timber
 
 @OptIn(ExperimentalPagingApi::class)
-class FollowersRemoteMediator internal constructor(
+class MutesListRemoteMediator(
     private val accountRepository: AccountRepository,
     private val databaseDelegate: DatabaseDelegate,
-    private val followersRepository: FollowersRepository,
     private val relationshipRepository: RelationshipRepository,
-    private val accountId: String,
-) : RemoteMediator<Int, FollowerWrapper>() {
+    private val mutesRepository: MutesRepository,
+) : RemoteMediator<Int, MuteWrapper>() {
     private var nextKey: String? = null
     private var nextPositionIndex = 0
 
     @Suppress("ReturnCount")
     override suspend fun load(
         loadType: LoadType,
-        state: PagingState<Int, FollowerWrapper>,
+        state: PagingState<Int, MuteWrapper>,
     ): MediatorResult {
         return try {
             var pageSize: Int = state.config.pageSize
@@ -37,10 +34,9 @@ class FollowersRemoteMediator internal constructor(
                 when (loadType) {
                     LoadType.REFRESH -> {
                         pageSize = state.config.initialLoadSize
-                        accountRepository.getAccountFollowers(
-                            accountId = accountId,
-                            olderThanId = null,
-                            loadSize = pageSize,
+                        mutesRepository.getMutes(
+                            maxId = null,
+                            limit = pageSize,
                         )
                     }
 
@@ -52,36 +48,32 @@ class FollowersRemoteMediator internal constructor(
                         if (nextKey == null) {
                             return MediatorResult.Success(endOfPaginationReached = true)
                         }
-                        accountRepository.getAccountFollowers(
-                            accountId = accountId,
-                            olderThanId = nextKey,
-                            loadSize = pageSize,
+                        mutesRepository.getMutes(
+                            maxId = nextKey,
+                            limit = pageSize,
                         )
                     }
                 }
 
-            val relationships = response.accounts.getRelationships(accountRepository)
+            val relationships =
+                accountRepository.getAccountRelationships(response.accounts.map { it.accountId })
 
             databaseDelegate.withTransaction {
                 if (loadType == LoadType.REFRESH) {
-                    followersRepository.deleteFollowers(accountId)
+                    mutesRepository.deleteAll()
                     nextPositionIndex = 0
                 }
 
                 accountRepository.insertAll(response.accounts)
                 relationshipRepository.insertAll(relationships)
-                followersRepository.insertAll(
+                mutesRepository.insertAll(
                     response.accounts.mapIndexed { index, account ->
-                        Follower(
-                            accountId = accountId,
-                            followerAccountId = account.accountId,
-                            position = nextPositionIndex + index,
-                        )
+                        account.toDatabaseMute(isMuted = true, position = index)
                     },
                 )
             }
 
-            nextKey = response.pagingLinks?.getMaxIdValue()
+            nextKey = response.nextPage?.maxId
             nextPositionIndex += response.accounts.size
 
             // There seems to be some race condition for refreshes.  Subsequent pages do
@@ -96,20 +88,20 @@ class FollowersRemoteMediator internal constructor(
             }
 
             @Suppress("KotlinConstantConditions")
-            MediatorResult.Success(
-                endOfPaginationReached =
-                    when (loadType) {
-                        LoadType.PREPEND -> response.pagingLinks?.find { it.rel == Rel.PREV } == null
-                        LoadType.REFRESH,
-                        LoadType.APPEND,
-                        -> response.pagingLinks?.find { it.rel == Rel.NEXT } == null
-                    },
-            )
+            (MediatorResult.Success(
+                endOfPaginationReached = when (loadType) {
+                    LoadType.PREPEND -> response.prevPage == null
+                    LoadType.REFRESH,
+                    LoadType.APPEND -> response.nextPage == null
+                },
+            ))
         } catch (e: Exception) {
             Timber.e(e)
             MediatorResult.Error(e)
         }
     }
-}
 
-private const val REFRESH_DELAY = 200L
+    companion object {
+        private const val REFRESH_DELAY = 200L
+    }
+}
