+++ feature/account/build.gradle.kts
@@ -11,6 +11,7 @@ dependencies {
     implementation(project(":core:datastore"))
     implementation(project(":core:usecase:mastodon"))
     implementation(project(":core:model"))
+    implementation(project(":core:repository:paging"))
     implementation(project(":core:repository:mastodon"))
     implementation(project(":core:designsystem"))
     implementation(project(":core:ui:common"))
