+++ feature/account/src/main/java/org/mozilla/social/feature/account/AccountModule.kt
@@ -8,6 +8,7 @@ import org.mozilla.social.core.analytics.analyticsModule
 import org.mozilla.social.core.datastore.dataStoreModule
 import org.mozilla.social.core.navigation.navigationModule
 import org.mozilla.social.core.repository.mastodon.mastodonRepositoryModule
+import org.mozilla.social.core.repository.paging.AccountTimelineRemoteMediator
 import org.mozilla.social.core.ui.postcard.postCardModule
 import org.mozilla.social.core.usecase.mastodon.mastodonUsecaseModule
 import org.mozilla.social.feature.account.edit.EditAccountViewModel
