+++ feature/favorites/src/main/java/org/mozilla/social/feature/favorites/FavoritesModule.kt
@@ -19,6 +19,5 @@ val favoritesModule = module {
         navigationModule,
     )
 
-    factoryOf(::FavoritesRemoteMediator)
     viewModelOf(::FavoritesViewModel)
 }
