+++ feature/favorites/src/main/java/org/mozilla/social/feature/favorites/FavoritesViewModel.kt
@@ -11,6 +11,7 @@ import org.koin.core.component.inject
 import org.koin.core.parameter.parametersOf
 import org.mozilla.social.core.analytics.AnalyticsIdentifiers
 import org.mozilla.social.core.repository.mastodon.FavoritesRepository
+import org.mozilla.social.core.repository.paging.FavoritesRemoteMediator
 import org.mozilla.social.core.ui.postcard.PostCardDelegate
 import org.mozilla.social.core.ui.postcard.toPostCardUiState
 import org.mozilla.social.core.usecase.mastodon.account.GetLoggedInUserAccountId
