+++ feature/feed/build.gradle.kts
@@ -16,6 +16,7 @@ dependencies {
     implementation(project(":core:designsystem"))
     implementation(project(":core:model"))
     implementation(project(":core:navigation"))
+    implementation(project(":core:repository:paging"))
     implementation(project(":core:repository:mastodon"))
     implementation(project(":core:ui:common"))
     implementation(project(":core:ui:postcard"))
