+++ feature/feed/src/main/java/org/mozilla/social/feed/FeedModule.kt
@@ -6,12 +6,12 @@ import org.mozilla.social.common.commonModule
 import org.mozilla.social.core.analytics.analyticsModule
 import org.mozilla.social.core.datastore.dataStoreModule
 import org.mozilla.social.core.navigation.navigationModule
-import org.mozilla.social.core.repository.mastodon.mastodonRepositoryModule
+import org.mozilla.social.core.repository.paging.FederatedTimelineRemoteMediator
+import org.mozilla.social.core.repository.paging.HomeTimelineRemoteMediator
+import org.mozilla.social.core.repository.paging.LocalTimelineRemoteMediator
+import org.mozilla.social.core.repository.paging.pagingModule
 import org.mozilla.social.core.ui.postcard.postCardModule
 import org.mozilla.social.core.usecase.mastodon.mastodonUsecaseModule
-import org.mozilla.social.feed.remoteMediators.FederatedTimelineRemoteMediator
-import org.mozilla.social.feed.remoteMediators.HomeTimelineRemoteMediator
-import org.mozilla.social.feed.remoteMediators.LocalTimelineRemoteMediator
 
 val feedModule =
     module {
@@ -19,7 +19,7 @@ val feedModule =
             commonModule,
             mastodonUsecaseModule,
             dataStoreModule,
-            mastodonRepositoryModule,
+            pagingModule,
             postCardModule,
             navigationModule,
             analyticsModule,
