+++ feature/followers/build.gradle.kts
@@ -12,6 +12,7 @@ dependencies {
     implementation(project(":core:designsystem"))
     implementation(project(":core:ui:common"))
     implementation(project(":core:ui:accountfollower"))
+    implementation(project(":core:repository:paging"))
     implementation(project(":core:repository:mastodon"))
     implementation(project(":core:common"))
     implementation(project(":core:usecase:mastodon"))
