+++ feature/settings/src/main/java/org/mozilla/social/feature/settings/contentpreferences/blockedusers/BlockedUsersViewModel.kt
@@ -19,7 +19,7 @@ import org.mozilla.social.core.ui.common.account.toggleablelist.ToggleableAccoun
 import org.mozilla.social.core.usecase.mastodon.account.BlockAccount
 import org.mozilla.social.core.usecase.mastodon.account.GetLoggedInUserAccountId
 import org.mozilla.social.core.usecase.mastodon.account.UnblockAccount
-import org.mozilla.social.core.usecase.mastodon.remotemediators.BlocksListRemoteMediator
+import org.mozilla.social.core.repository.paging.BlocksListRemoteMediator
 import org.mozilla.social.feature.settings.R
 import org.mozilla.social.feature.settings.SettingsInteractions
