+++ feature/settings/src/main/java/org/mozilla/social/feature/settings/contentpreferences/mutedusers/MutedUsersSettingsViewModel.kt
@@ -19,7 +19,7 @@ import org.mozilla.social.core.ui.common.account.toggleablelist.ToggleableAccoun
 import org.mozilla.social.core.usecase.mastodon.account.GetLoggedInUserAccountId
 import org.mozilla.social.core.usecase.mastodon.account.MuteAccount
 import org.mozilla.social.core.usecase.mastodon.account.UnmuteAccount
-import org.mozilla.social.core.usecase.mastodon.remotemediators.MutesListRemoteMediator
+import org.mozilla.social.core.repository.paging.MutesListRemoteMediator
 import org.mozilla.social.feature.settings.R
 import org.mozilla.social.feature.settings.SettingsInteractions
