+++ settings.gradle.kts
@@ -39,7 +39,9 @@ include(":core:network:mastodon")
 include(":core:network:mozilla")
 include(":core:repository:mastodon")
 include(":core:repository:mozilla")
+include(":core:repository:paging")
 include(":core:test")
+include(":core:ui:accountfollower")
 include(":core:ui:common")
 include(":core:ui:htmlcontent")
 include(":core:ui:poll")
@@ -57,5 +59,4 @@ include(":feature:report")
 include(":feature:hashtag")
 include(":feature:followers")
 include(":feature:discover")
-include(":core:ui:accountfollower")
 include(":feature:favorites")
